Added open all sites button to vote interface
It sometimes occurs that a captcha popup is not clearly accessible in the VoteAssist iFrame. Because of that I added an option to open all registered vote sites each in a separate tab. The user will probably have to allow popups for the gforum.co domain, though this setting will be saved